feat: 용자가 작성한 동행찾기 게시글 목록 반환값 추가 - status, createdAt
